dbGap data download for a breakseq test.
Only the history of how was it made included, since data directory
is not tracked. Fastq output was exported directly to Crick, where
data can be found. SRA files and ngc key are in my local directory.

There is also a branch with the tests that I used to come up with
this methodology in this same GitHub repository.