Code to download dbGap data
Now that we know that breakseq will work, I have semi-automatized
the dbGap download process. It is not fully automatized (via a loop
over the accessions list) because there is not enough space to
store so much data (each file is roughly 200Gb).

For an easy follow-up of the location of each file, the accession
list is modified both in local and Crick each time there is an up-
date in the download process.