Imputation run
The imputation process was re-run due to minor changes in the data:
when transforming Bell et al. 2019 data into hg19 an additional
step checked that reference and alternative alleles were the same
as in the 1000 genomes project data.

This data was used in the new PCA made with all the autosomes so this
imputations was run for consistency and to have it automatized in the
new file system.