Minor adjustement to avoid errors when filtering unknown genotypes.
I noticed that this code would delete not only lines that have
unknown genotypes but also those that have individual names that
cointain 'Del' or '.' due to the use of `grep` instead of `awk.

This did not have an effect in the present imputation results, but
would potentially affect imputation attempts with other inversion
genotypes.

This priciple will also be applied to the code to check tag SNPs.