Run Plink to check tag SNPs in a set of individuals and inversions
This code is very similar to the imputation one, in fact, it would
be more optimal to have both processes run together. However, since
I want the different parts of the code to be able to run independently,
it was made this way.

As it can be seen in immediately previous commits, this process
allowed me to improve the imputation process too.

As explained in the history file, there was a small mistake in the
code that concatenates all the ld process outputs so I solved it and
ran only the interesting part again.

As an advantage of this re-running, in the logfile corresponding to
2020-04-27 there is a list of those inversions and populations that
did not have enough variability in the population to be analyzed. In
other words, they are monomorphic for these 20 individuals.

This leads me to think that maybe I should re-run this process with
both 1000 genomes and my 20 individuals, at least for these inversions.
However, for the rest of them, it is enough to search for the reported
tag SNPs and check their LD with their corresponding inversion in these
 20 individuals, which will probably be the next step.