New tagsnp runs with failed inversions and SAS
These runs have in common that individuals from Bell et al. 2019
were analyzed together with 1000GP individuals.

Some inversions failed the last time for AFR and EAS because there
are not much individuals. Those inversions were repeated and results
are stored in the directory with the `_failedinvs` tag.

Also, since there is only one SAS individual, all inversions were
analyzed in this population together with 1000 GP from SAS population
and the results are stored in the directory with the `_SAS` tag.

Some changes in the code were applied for it to be adapted to process
directories with either one file with all chromosomes or multiple
files, one for each chromosome.