Solved problem with tagSNP checking
There were two main issues in previous tag SNP checking procedures:

1. Too strict genotyping filter
Only tagSNPs from genotyped individuals can be checked. By being too
strict with the genotype filters, a lot of tagSNPs were not being
analyzed. The new version for `runImputation.sh` allows for almost
any genotype that is not right away wrong.

2. Bad individual ID grep
Names in our individuals can easily confuse processes like `grep`,
because there is an nc2, nc20, nc21, nc22 and so on. All of them are
found when grepping for "nc2". To solve that, I made `grep` in
`checkTagSNPs.sh` (and in `runImputation.sh` as well, because they
share a fair bit of code) to search for the individual + spacer
character.