Incorporated to runall.sh new crossover data processing.
This part of the code is to improve the resolution in crossovers
data. The mean resolution is about 400kb, but smaller  patterns
were observed when overlaping all crossover events together.

To mimic this numerically, this part of the code makes windows of
desired size and then sums the fractions of crossover tracks that
overlap to each window.

It also creates a table to know which inversions overlap with which
windows.

Later on the analysis, this has to be normalized by number of
samples.