New python script makeBedWindows
As an effort to migrate from Bash to Python as the main used language,
an important part of runall.sh was translated into a Python script.

makeBedWindows recieves a list of target regions and a series of
windowing instructions via arguments, and creates the desired windows
in bed format. It is perfect to make targeted windows to use later with
bedtools.